Integrate symbol table tests into existing infrastructure
